Refactor CMake and QMake targets structuring and building
* For CMake, all targets which were compiled as static libraries, now
will be compiled as objects, thus making EXPORT_IMPORT_MAGIC redundant.
Therefore, file globals.h was deleted
* Target dependency for CMake build was reviewed and adjusted in a way
that only Qt modules will be available as PUBLIC, so that they will be
available implicitly. Everything else will be linked as PRIVATE, meaning
that "if you want to use something - link against it, do not wait that
target would be accessible implicitly".
* QML types are now separated as external (resources which are
explicitly related to Jira) and internal (resources which are needed
internally).
* For QMake, plugin as a shared library is not used anymore, instead of
that - all cpp tests are now including sources of units which they are
testing. This unifies CMake and QMake testing approach - "do not test
library, test units instead!"
* Unused #include headers were removed