Redesign Session test case
Public API of Session cannot be verified without validation of Reply
status code, therefore after sending GET or POST request, a status code
of Reply will be used to check that GET and POST methods were really
used.
In addition to that, test case was moved to CuteMockServer so that there
is no dependency on remote hosts anymore.