Use temporary QML objects in tests
Previously objects were not destroyed after failing/passing tests. Even
though life time and memory consumption of QML test cases was not high,
it is a good practice to not forget that when test is failing and object
is not released, it might cause some issues with shared resources in the
follow-up tests.