Remove network tests related to Host not found error
Removed "Host not found" data for network error tests in order to
speedup auto tests on Windows platform. Also, several tests were
receiving status 404 instead of 503 due to incorrect recognition of
routes, but they did not fail since plugin was not (and should not)
recognizing error codes for network issues - this issue was fixed.