Reduce coupling for derived endpoints
Instance of Jira class will be always assigned as a parent for any given
derived endpoint. Instance of an active Session in a nutshell provides a
shared QNetworkAccessManager with Cookie Jar, so that any endpoint will
be represented by one single session. Design of Jira class makes Session
private, therefore an initial design introduced unnecessary coupling by
making a requirement to pass an active Session to endpoint constructor.
As a solution which would enable testing of a generic endpoint and
reduce coupling with derived endpoints, an Endpoint class has been made
a friend class of Jira.